--- mex/seq_repo/models/__init__.py
@@ -1,3 +1,0 @@
-from mex.seq_repo.models.source import SeqRepoSource
-
-__all__ = ("SeqRepoSource",)
